modify batch compare scripts

+++ CompareWithBaseline/RTDA2CoresPerformance/CompareRTDA2Cores.sh
@@ -4,34 +4,41 @@
 # ************** Adjust settings there **************
 title="RTDA2CoresPerformance"
 MinTaskNumber=3
-MaxTaskNumber=20
+MaxTaskNumber=5
 ## no separator '/' at the end of the path
 #ROOT_PATH="/home/zephyr/Programming/DAG_NLP" 
 ROOT_PATH="/home/dong/workspace/DAG_NLP"
 RESULTS_PATH="$ROOT_PATH/TaskData/dagTasks"
-methods_dir_name=( "Initial_Res" "OptOrder_Res" "Verucchi_Res" "NLP_Res" )
-makeProgressTimeLimit=20
-kVerucchiTimeLimit=20
+# methods_dir_name=( "Initial_Res" "OptOrder_Res" "Verucchi_Res" "NLP_Res" )
+methods_dir_name=( "Initial_Res" "OptOrder_Res" "Verucchi_Res" )
+makeProgressTimeLimit=10
+kVerucchiTimeLimit=10
 coreNumberAva=2
-keep_current_result=False
+keep_current_result_and_only_plot=0
+history_result_directory="$ROOT_PATH/CompareWithBaseline/RTDA2CoresPerformance"
 ## setting for generating task sets
 TaskSetType=1
-taskSetNumber=3
+taskSetNumber=2
 randomSeed=-1 # negative means time seed
 # ***************************************************
 # ***************************************************
 
-if [[ $keep_current_result == 0 || $keep_current_result == false || $keep_current_result == False ]]; then
-  echo "Clearing all current results."
-  for dir_name in ${methods_dir_name[@]}; do
-    if [[ -d $dir_name ]]; then rm -rf $dir_name; fi
-    mkdir $dir_name
-  done
-else
-  for dir_name in ${methods_dir_name[@]}; do
-    if [[ ! -d $dir_name ]]; then mkdir $dir_name; fi
-  done
+if [[ $keep_current_result_and_only_plot == 1 || $keep_current_result_and_only_plot == true \
+  || $keep_current_result_and_only_plot == True || $keep_current_result_and_only_plot == TRUE ]]; then
+  echo "Plot from history in directory: $history_result_directory"
+  # visualize history result
+  python $ROOT_PATH/CompareWithBaseline/$title/Visualize_RTDA_performance.py --minTaskNumber $MinTaskNumber \
+    --title $title --maxTaskNumber $MaxTaskNumber --result_file_path $history_result_directory
+  exit
 fi
+
+echo "Clearing all current results."
+for dir_name in ${methods_dir_name[@]}; do
+  if [[ -d $dir_name ]]; then rm -rf $dir_name; fi
+  mkdir $dir_name
+done
+if [[ -d dagTasks ]]; then rm -rf dagTasks; fi
+  mkdir dagTasks
 
 cp parameters.yaml $ROOT_PATH/sources/parameters.yaml
 # build the project
@@ -44,12 +51,12 @@ python $ROOT_PATH/CompareWithBaseline/edit_yaml.py --entry "makeProgressTimeLimi
 python $ROOT_PATH/CompareWithBaseline/edit_yaml.py --entry "kVerucchiTimeLimit" --value $kVerucchiTimeLimit
 
 perform_optimization() {
-# Optimize energy consumption
-cd $ROOT_PATH/release
-make -j8
-./tests/tBatch1
-cd $ROOT_PATH/CompareWithBaseline/$title
-sleep 1
+  # Optimize energy consumption
+  cd $ROOT_PATH/release
+  cmake --build . --config Release -- -j 8
+  ./tests/tBatch1
+  cd $ROOT_PATH/CompareWithBaseline/$title
+  sleep 1
 }
 
 for (( jobNumber=$MinTaskNumber; jobNumber<=$MaxTaskNumber; jobNumber++ ))
@@ -62,9 +69,9 @@ do
 	python $ROOT_PATH/CompareWithBaseline/edit_yaml.py --entry "weightSF_factor" --value 0
 	perform_optimization
 
-	# wang_21rtssIC
-	python $ROOT_PATH/CompareWithBaseline/edit_yaml.py --entry "weightSF_factor" --value 1
-	perform_optimization
+	# # wang_21rtssIC
+	# python $ROOT_PATH/CompareWithBaseline/edit_yaml.py --entry "weightSF_factor" --value 1
+	# perform_optimization
 	
   #copy results to corresponding folder
 	taskset_folder_name="N$jobNumber"
@@ -76,8 +83,11 @@ do
     cat ./$taskset_folder_name/* >> $taskset_result_summary_file_name
     cd $ROOT_PATH/CompareWithBaseline/$title
   done
+  if [[ ! -d dagTasks/$taskset_folder_name ]]; then mkdir dagTasks/$taskset_folder_name; fi
+  cp $RESULTS_PATH/*.csv dagTasks/$taskset_folder_name/
 
 done
 
-visualize the result
-python $ROOT_PATH/CompareWithBaseline/$title/Visualize_performance.py  --minTaskNumber $MinTaskNumber --title $title  --maxTaskNumber $MaxTaskNumber --root_path $ROOT_PATH
+# visualize the result
+python $ROOT_PATH/CompareWithBaseline/$title/Visualize_RTDA_performance.py --minTaskNumber $MinTaskNumber \
+  --title $title --maxTaskNumber $MaxTaskNumber --result_file_path $ROOT_PATH/CompareWithBaseline/$title
